+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `IronTrail::DbFunctions#trigger_errors_metrics` method to be able to monitor error metrics
+
 ## 0.1.0 - 2025-01-29
 
 ### Changed